fixed issues reported by Kyle
There were some issues with the list that I didn't notice during my initial testing.  When it changed to a multi-dimensional list, the program no longer worked correctly, the code has been modified and a for loop has been added to the saveTestToDB().  I had previously had ran this method once then had to remove it so duplicate tests would not be added.  I have made this much cleaner.  The saveTestToDB() method will only run if no test data is found in the db.json file.  An if statement at the beginning of this method accomplishes this.